// frontend/lib/api-client/models/PaginatedTarefaListList.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { TarefaList } from './TarefaList';
export type PaginatedTarefaListList = {
    count?: number;
    next?: string | null;
    previous?: string | null;
    results?: Array<TarefaList>;
};